// lib/tdlib/src/generated/td-api.ts
export interface TdOk {
  readonly '@type': 'ok';
}
// ...
export interface TdAuthenticationCodeInfo {
  readonly '@type': 'authenticationCodeInfo';
  phone_number: string;
  type: TdAuthenticationCodeType | null;
  next_type: TdAuthenticationCodeType | null;
  timeout: number;
}

export interface TdEmailAddressAuthenticationCodeInfo {
  readonly '@type': 'emailAddressAuthenticationCodeInfo';
  email_address_pattern: string;
  length: number;
}
// ...
export interface TdTermsOfService {
  readonly '@type': 'termsOfService';
  text: TdFormattedText | null;
  min_user_age: number;
  show_popup: boolean;
}

export interface TdPasskey {
  readonly '@type': 'passkey';
  id: string;
  name: string;
  addition_date: number;
  last_usage_date: number;
  software_icon_custom_emoji_id: number;
}

export interface TdPasskeys {
  readonly '@type': 'passkeys';
  passkeys: TdVector<TdPasskey>;
}

export interface TdAuthorizationStateWaitTdlibParameters {
  readonly '@type': 'authorizationStateWaitTdlibParameters';
}

export interface TdAuthorizationStateWaitPhoneNumber {
  readonly '@type': 'authorizationStateWaitPhoneNumber';
}

export interface TdAuthorizationStateWaitPremiumPurchase {
  readonly '@type': 'authorizationStateWaitPremiumPurchase';
  store_product_id: string;
  premium_day_count: number;
  support_email_address: string;
  support_email_subject: string;
}

export interface TdAuthorizationStateWaitEmailAddress {
  readonly '@type': 'authorizationStateWaitEmailAddress';
  allow_apple_id: boolean;
  allow_google_id: boolean;
}

export interface TdAuthorizationStateWaitEmailCode {
  readonly '@type': 'authorizationStateWaitEmailCode';
  allow_apple_id: boolean;
  allow_google_id: boolean;
  code_info: TdEmailAddressAuthenticationCodeInfo | null;
  email_address_reset_state: TdEmailAddressResetState | null;
}

export interface TdAuthorizationStateWaitCode {
  readonly '@type': 'authorizationStateWaitCode';
  code_info: TdAuthenticationCodeInfo | null;
}

export interface TdAuthorizationStateWaitOtherDeviceConfirmation {
  readonly '@type': 'authorizationStateWaitOtherDeviceConfirmation';
  link: string;
}

export interface TdAuthorizationStateWaitRegistration {
  readonly '@type': 'authorizationStateWaitRegistration';
  terms_of_service: TdTermsOfService | null;
}

export interface TdAuthorizationStateWaitPassword {
  readonly '@type': 'authorizationStateWaitPassword';
  password_hint: string;
  has_recovery_email_address: boolean;
  has_passport_data: boolean;
  recovery_email_address_pattern: string;
}

export interface TdAuthorizationStateReady {
  readonly '@type': 'authorizationStateReady';
}

export interface TdAuthorizationStateLoggingOut {
  readonly '@type': 'authorizationStateLoggingOut';
}

export interface TdAuthorizationStateClosing {
  readonly '@type': 'authorizationStateClosing';
}

export interface TdAuthorizationStateClosed {
  readonly '@type': 'authorizationStateClosed';
}
// ...
export type TdAuthorizationState =
  | TdAuthorizationStateWaitTdlibParameters
  | TdAuthorizationStateWaitPhoneNumber
  | TdAuthorizationStateWaitPremiumPurchase
  | TdAuthorizationStateWaitEmailAddress
  | TdAuthorizationStateWaitEmailCode
  | TdAuthorizationStateWaitCode
  | TdAuthorizationStateWaitOtherDeviceConfirmation
  | TdAuthorizationStateWaitRegistration
  | TdAuthorizationStateWaitPassword
  | TdAuthorizationStateReady
  | TdAuthorizationStateLoggingOut
  | TdAuthorizationStateClosing
  | TdAuthorizationStateClosed;
// ...
export type TdEmailAddressResetState =
  | TdEmailAddressResetStateAvailable
  | TdEmailAddressResetStatePending;
// ...
export interface TdClose {
  readonly '@type': 'close';
}
// ...
export type TdResultFor<T extends TdRequest> = T['@type'] extends keyof TdResultMap
  ? TdResultMap[T['@type']]
  : never;